refactor: mannerPoints 기능 분리
